requirements.txt
openai
qdrant-client
gradio
beautifulsoup4
python-dotenv
